+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.25.05
 
+### Added
+* ***vibrational***: New function interface `full_boltz_pop` to compute the analytic total Boltzmann population for a given set of frequencies at any temperature.
+
 ### Fixed
 * ***input***: Fixed parsing of Gaussian version in fchk files generated from checkpoint files obtained with the `c86dv` (form: CDVRev-X.XX).
 * ***vibrational***: Fixed missing initialization of `nqi` in `boltz_pop_max_quanta` when checking duplicate modes, which could lead to the false assertion that the list of mode indexes contained duplicates.